global_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.2
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fonttools==4.60.1
frozenlist==1.7.0
gevent==25.8.2
greenlet==3.2.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
multidict==6.6.4
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pybit==5.11.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
pyTelegramBotAPI==4.29.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
regex==2025.9.1
requests==2.32.5
schedule==1.2.2
six==1.17.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
